+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.8] - 2023-02-17
+
+### Changed
+* Automate milestone XP gain
+
 ### Fixed
 * Ensure that modifier penalties and modifiers are always positive